Refactor models module
