+++ back/tmeit_backend/models/__init__.py
@@ -1,1 +1,2 @@
-from .members.members import Member, Base
+from .members.members import Member
+from .sign_up import SignUp
